mf-karr-backend/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0 
flask_cors>=5.0.0
requests==2.31.0
psycopg2-binary==2.9.10
pycryptodome>=3.22.0